# docs/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QualiDiagIQ API",
    "description": "\n        Motor de Diagnóstico Tributário Automatizado para a Reforma do Consumo.\n        \n        Parte do ecossistema Tributiq. Fornece análise de maturidade e conformidade\n        com base na EC 132/2023, LC 214/2025 e ABNT NBR 17301:2026.\n        ",
    "contact": {
      "name": "Equipe Tributiq",
      "email": "contato@tributiq.com.br"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/diagnosticos/rascunho-self-service": {
      "post": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Gravar rascunho do diagnóstico na base (antes do OTP)",
        "description": "Persiste o mesmo corpo de **POST /diagnosticos/self-service** na tabela de rascunhos e envia OTP. O cliente deve guardar apenas o **resgate_token** (fragmento de URL ou memória de curto prazo) — não usar o payload completo em sessionStorage como etapa final. **Idempotency-Key** obrigatório (middleware).",
        "operationId": "criar_rascunho_diagnostico_self_service_diagnosticos_rascunho_self_service_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IniciarDiagnosticoRequest"
              },
              "examples": {
                "micro_comercio_um_item": {
                  "summary": "Micro comércio — uma resposta escala",
                  "description": "Headers obrigatorios: Authorization Bearer (JWT valido do login) e Idempotency-Key (UUID v4). Resposta HTTP 201 com score e consultoria derivada.",
                  "value": {
                    "empresa": {
                      "cnpj": "12345678000195",
                      "razao_social": "Empresa Exemplo LTDA",
                      "porte": "micro",
                      "regime": "simples_nacional",
                      "cnae_principal": "1234567",
                      "uf": "SP",
                      "setor_macro": "comercio"
                    },
                    "respondente": {
                      "email": "fiscal@empresa.com.br",
                      "nome": "Fulano da Silva",
                      "telefone": "11999998888"
                    },
                    "respostas": [
                      {
                        "pergunta_id": "df52b20e-dab8-5ce5-a89d-4fb235016cbe",
                        "valor": 4
                      }
                    ],
                    "plano": "gratuito",
                    "locale_relatorio": "pt-BR",
                    "aceite_termos_privacidade": true
                  }
                },
                "servicos_medio_presumido": {
                  "summary": "Serviços — médio porte, Lucro Presumido",
                  "description": "Mesmo requisito de cadastro PJ: CNPJ válido (14 dígitos) + razão social e perfil fiscal.",
                  "value": {
                    "empresa": {
                      "cnpj": "11222333000181",
                      "razao_social": "Empresa Anônima Exemplo LTDA",
                      "porte": "medio",
                      "regime": "lucro_presumido",
                      "cnae_principal": "6201501",
                      "uf": "SP",
                      "setor_macro": "servicos"
                    },
                    "respondente": {
                      "email": "contador@escritorio.com.br",
                      "nome": "Maria Souza"
                    },
                    "respostas": [
                      {
                        "pergunta_id": "df52b20e-dab8-5ce5-a89d-4fb235016cbe",
                        "valor": 3
                      }
                    ],
                    "plano": "gratuito",
                    "locale_relatorio": "pt-BR",
                    "aceite_termos_privacidade": true
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RascunhoDiagnosticoSelfServiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/diagnosticos/rascunho-self-service/resumo": {
      "get": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Resumo do rascunho (token opaco)",
        "description": "Metadados mínimos para a página de confirmação. **Header obrigatório:** `X-Rascunho-Token` com o valor devolvido por POST /diagnosticos/rascunho-self-service.",
        "operationId": "resumo_rascunho_diagnostico_self_service_diagnosticos_rascunho_self_service_resumo_get",
        "parameters": [
          {
            "name": "X-Rascunho-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Rascunho-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticoRascunhoResumoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/diagnosticos/rascunho-self-service/concluir": {
      "post": {
        "tags": [
          "Diagnósticos"
        ],
        "summary": "Concluir rascunho com OTP (grava diagnóstico final)",
        "description": "Valida o código enviado por e-mail, materializa **POST /diagnosticos/self-service** a partir do JSON guardado no rascunho e marca o rascunho como consumido. **Idempotency-Key** obrigatório.",
        "operationId": "concluir_rascunho_diagnostico_self_service_diagnosticos_rascunho_self_service_concluir_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcluirRascunhoDiagnosticoSelfServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConcluirRascunhoDiagnosticoSelfServiceRequest": {
        "properties": {
          "resgate_token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 16,
            "title": "Resgate Token"
          },
          "codigo": {
            "type": "string",
            "maxLength": 16,
            "minLength": 4,
            "title": "Codigo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resgate_token",
          "codigo"
        ],
        "title": "ConcluirRascunhoDiagnosticoSelfServiceRequest",
        "description": "OTP + token de resgate do rascunho persistido na BD."
      },
...
      "DiagnosticoRascunhoResumoResponse": {
        "properties": {
          "empresa_razao_social": {
            "type": "string",
            "title": "Empresa Razao Social"
          },
          "empresa_cnpj": {
            "type": "string",
            "title": "Empresa Cnpj",
            "description": "CNPJ normalizado no payload do rascunho (14 dígitos) ou vazio — UX antes de vincular à conta.",
            "default": ""
          },
          "email_mascarado": {
            "type": "string",
            "title": "Email Mascarado"
          },
          "respondente_email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 3,
            "title": "Respondente Email",
            "description": "E-mail normalizado gravado no rascunho (mesmo usado em POST)."
          },
          "expira_em": {
            "type": "string",
            "format": "date-time",
            "title": "Expira Em"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "empresa_razao_social",
          "email_mascarado",
          "respondente_email",
          "expira_em"
        ],
        "title": "DiagnosticoRascunhoResumoResponse",
        "description": "Metadados para a página de OTP.\n\nO portador do ``X-Rascunho-Token`` já provou posse do fluxo; incluímos o e-mail em claro para\nreenvio de código (paridade com /auth/verificar-email/solicitar)."
      },
      "DiagnosticoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "plano": {
            "type": "string",
            "title": "Plano"
          },
          "empresa_razao_social": {
            "type": "string",
            "title": "Empresa Razao Social"
          },
          "empresa_cnpj": {
            "type": "string",
            "title": "Empresa Cnpj",
            "description": "CNPJ no snapshot do diagnóstico (14 dígitos ou vazio se não informado).",
            "default": ""
          },
          "criado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Em",
            "description": "Instante de criação (UTC) — útil para linha do tempo e comparativo entre ciclos."
          },
          "finalizado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizado Em",
            "description": "Instante de finalização (UTC), se aplicável."
          },
          "empresa_faixa_faturamento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empresa Faixa Faturamento",
            "description": "Faixa de faturamento autodeclarada no POST (slug canónico), se informada."
          },
          "locale_relatorio": {
            "type": "string",
            "title": "Locale Relatorio",
            "description": "Idioma usado na geração do PDF (persistido com o diagnóstico).",
            "default": "pt-BR"
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreCompletoSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "relatorio_pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relatorio Pdf Url"
          },
          "recomendacao_ia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recomendacao Ia"
          },
          "checklist": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist"
          },
          "matriz_impacto": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matriz Impacto"
          },
          "cronograma": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cronograma"
          },
          "checklist_m12_autoconf": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist M12 Autoconf"
          },
          "quadro_implantacao_anotacoes": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quadro Implantacao Anotacoes",
            "description": "Mapa f{i}_a{j} -> {prazo_meta, comentarios[]} (legado comentario migrado na leitura)."
          },
          "aceite_termos_privacidade_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aceite Termos Privacidade Em"
          },
          "hash_evidencia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hash Evidencia"
          },
          "versao_otimista": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versao Otimista"
          },
          "versao_plano": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao Plano",
            "description": "Versão do snapshot do plano de ação materializado (checklist/matriz/cronograma).",
            "default": 1
          },
          "leitura_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leitura Token",
            "description": "Presente só em POST /diagnosticos/rascunho-self-service/concluir: token opaco para GET /diagnosticos/self-service/conclusao-visualizacao (leitura pública limitada no tempo)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "plano",
          "empresa_razao_social"
        ],
        "title": "DiagnosticoResponse"
      },
...
      "EmpresaSchema": {
        "properties": {
          "cnpj": {
            "type": "string",
            "maxLength": 18,
            "title": "Cnpj",
            "description": "CNPJ: **opcional** em fluxos só self-service (rascunho / OTP sem conta na plataforma). Vazio = sem cadastro PJ naquele ciclo. Se informado: 14 dígitos ou máscara, DV RFB válido. Com **sessão na plataforma** ou **vinculação de rascunho à conta**, o CNPJ é **obrigatório** (histórico por empresa no tenant) — ver `EmpresaPainelSchema` / ADR-013.",
            "default": ""
          },
          "razao_social": {
            "type": "string",
            "title": "Razao Social"
          },
          "porte": {
            "$ref": "#/components/schemas/PorteEmpresa"
          },
          "regime": {
            "$ref": "#/components/schemas/RegimeTributario"
          },
          "cnae_principal": {
            "type": "string",
            "maxLength": 7,
            "minLength": 7,
            "title": "Cnae Principal",
            "description": "CNAE principal com 7 dígitos"
          },
          "uf": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Uf",
            "description": "Sigla da UF com 2 letras"
          },
          "setor_macro": {
            "$ref": "#/components/schemas/SetorMacro"
          },
          "faixa_faturamento": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FaixaFaturamentoDeclarada"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opcional — faixa de faturamento bruto anual autodeclarada (R$), para segmentação; omitir ou null se o respondente não informar."
          }
        },
        "type": "object",
        "required": [
          "razao_social",
          "porte",
          "regime",
          "cnae_principal",
          "uf",
          "setor_macro"
        ],
        "title": "EmpresaSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IniciarDiagnosticoRequest": {
        "properties": {
          "empresa": {
            "$ref": "#/components/schemas/EmpresaSchema"
          },
          "respondente": {
            "$ref": "#/components/schemas/RespondenteSchema"
          },
          "respostas": {
            "items": {
              "$ref": "#/components/schemas/RespostaRequestSchema"
            },
            "type": "array",
            "title": "Respostas"
          },
          "plano": {
            "type": "string",
            "title": "Plano",
            "default": "gratuito"
          },
          "locale_relatorio": {
            "type": "string",
            "enum": [
              "pt-BR",
              "en"
            ],
            "title": "Locale Relatorio",
            "description": "Idioma do relatório PDF (WeasyPrint): pt-BR completo; en — labels em inglês (conteúdo gerado dinamicamente pode permanecer em PT até tradução total).",
            "default": "pt-BR"
          },
          "aceite_termos_privacidade": {
            "type": "boolean",
            "title": "Aceite Termos Privacidade",
            "description": "Ciência explícita dos termos de uso e desta política de privacidade (LGPD Lei 13.709/2018)."
          },
          "force_refresh_cnpj": {
            "type": "boolean",
            "title": "Force Refresh Cnpj",
            "description": "Se verdadeiro e ``empresa.cnpj`` preenchido: nova consulta às fontes públicas ignorando TTL antes de fechar evidência WORM; merge com histórico em ``diagnostico_empresa_campo_historico``.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "empresa",
          "respondente",
          "respostas",
          "aceite_termos_privacidade"
        ],
        "title": "IniciarDiagnosticoRequest"
      },
...
      "PorteEmpresa": {
        "type": "string",
        "enum": [
          "micro",
          "pequeno",
          "medio",
          "grande"
        ],
        "title": "PorteEmpresa",
        "description": "Faixas de porte alinhadas com a segmentação da PwC Brasil\n(estudo \"Tributos no Centro: Caminhos para a Reinvenção\", nov/2025)."
      },
...
      "RascunhoDiagnosticoSelfServiceResponse": {
        "properties": {
          "resgate_token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 16,
            "title": "Resgate Token"
          },
          "mensagem": {
            "type": "string",
            "title": "Mensagem"
          },
          "expira_em": {
            "type": "string",
            "format": "date-time",
            "title": "Expira Em"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resgate_token",
          "mensagem",
          "expira_em"
        ],
        "title": "RascunhoDiagnosticoSelfServiceResponse",
        "description": "Resposta ao gravar rascunho no servidor (token opaco devolvido uma vez)."
      },
      "RegimeTributario": {
        "type": "string",
        "enum": [
          "simples_nacional",
          "lucro_presumido",
          "lucro_real",
          "mei"
        ],
        "title": "RegimeTributario",
        "description": "Regime tributário declarado pela empresa-cliente."
      },
...
      "RespondenteSchema": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome",
            "description": "Nome do respondente que vincula o diagnóstico ao contato (obrigatório)."
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 15
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone",
            "description": "Telefone opcional, apenas dígitos: DDD + número (10 ou 11 dígitos), sem DDI (+55). M09 lead na plataforma; LGPD por finalidade."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "nome"
        ],
        "title": "RespondenteSchema",
        "description": "Respondente do diagnóstico — e-mail e nome obrigatórios (vínculo LGPD / rastreio)."
      },
      "RespostaRequestSchema": {
        "properties": {
          "pergunta_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pergunta Id"
          },
          "valor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Valor"
          }
        },
        "type": "object",
        "required": [
          "pergunta_id",
          "valor"
        ],
        "title": "RespostaRequestSchema"
      },
      "ScoreCompletoSchema": {
        "properties": {
          "score_geral": {
            "$ref": "#/components/schemas/ScoreDimensaoSchema"
          },
          "score_por_dimensao": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ScoreDimensaoSchema"
            },
            "type": "object",
            "title": "Score Por Dimensao"
          }
        },
        "type": "object",
        "required": [
          "score_geral",
          "score_por_dimensao"
        ],
        "title": "ScoreCompletoSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}